Update Scores and User Turns on Webpage
Servlet getupdate request, gets the current user score and the username to who's turn it is. It passes this to webpage in format "##_xxxx". If it is the current users turn, the webpage says You!.

Also added a flag that blocks updates when playing a word to hopefully reduce unregistered button clicks. Seems a little better.